Improve inventory tests; few fixes (#9902)
* Add Inventory test class, factorize

* Add tests on drives; fixes

* Add tests on antivirus; fixes

* Add tests on batteries, fixes

* Add tests on bios

* Add tests on processors

* Add tests on cameras

* Add tests on controllers

* Adjust databaseinstance

* Add tests on firmwares

* Add tests on graphic cards

* Add tests on memories; fixes

* Add tests on networkcard

Drop useless overrided import mothod; adjust import cirteria

* Add tests on sensors

* Add tests on sound cards

* Add tests on drives and hard drives

* Typo

* Add tests on operating system

* Add tests on peripherals, fixes

* Add tests on remote managements; fixes

* Do not remove devices themselves; only links

* Order is not relevant

* Add tests on software

* Apply suggestions from code review


* Try to better handle entity in session

Revert not needed changes

* Restore all entites variables in session

* Apply suggestions from code review


* Handle correctly computer_items

Printers, Monitors and Peripherals must
move if inventoried from another computer

* Suggestions from code review

+++ inc/inventory/asset/device.class.php
@@ -64,17 +64,17 @@ abstract class Device extends InventoryAsset
       $iterator = $DB->request([
          'SELECT'    => [
             "$itemdevicetable.$fk",
+            "is_dynamic"
          ],
          'FROM'      => $itemdevicetable,
          'WHERE'     => [
             "$itemdevicetable.items_id"     => $this->item->fields['id'],
-            "$itemdevicetable.itemtype"     => $this->item->getType(),
-            "$itemdevicetable.is_dynamic"   => 1
+            "$itemdevicetable.itemtype"     => $this->item->getType()
          ]
       ]);
 
       foreach ($iterator as $row) {
-         $db_existing[$row[$fk]][] = $row[$fk];
+         $db_existing[$row[$fk]] = $row;
       }
 
       return $db_existing;
@@ -115,8 +115,6 @@ abstract class Device extends InventoryAsset
                      $fk => $device_id,
                      'items_id'     => $this->item->fields['id'],
                      'itemtype'     => $this->item->getType(),
-                     'is_dynamic'   => 1
-
                   ]
                );
                $deleted_items[$device_id] = $device_id;
@@ -133,15 +131,15 @@ abstract class Device extends InventoryAsset
             unset($existing[$device_id]);
          }
 
-         foreach (array_keys($existing) as $deviceid) {
+         foreach ($existing as $deviceid => $data) {
             //first, remove items
-            $DB->delete(
-               $itemdevice->getTable(), [
-                  $fk => $deviceid
-               ]
-            );
-            //then device itself
-            $device->delete(['id' => $deviceid], true);
+            if ($data['is_dynamic'] == 1) {
+               $DB->delete(
+                  $itemdevice->getTable(), [
+                     $fk => $deviceid
+                  ]
+               );
+            }
          }
       }
    }
